backend/requirements.txt
redis
uvicorn
fastapi
fastapi_limiter
pillow
python-dotenv
tqdm
numpy
pymongo
joblib
bcrypt
pyjwt
rich
mongodb-odm
google-genai
openai
partialjson
pyrate-limiter
ftfy
regex
git+https://github.com/openai/CLIP.git
webp
transformers[torch]
sentencepiece
protobuf
python-multipart
opencv-python
ultralytics
deepface
tf-keras
timm
PyNaCl
